Remove some possible race conditions in SLAF. Addresses #1545 (#1557)
* Remove some possible race conditions in SLAF. Addresses #1545

* Ulf Andrae: Update changelog

---------

+++ deode/mars_utils.py
@@ -146,7 +146,7 @@ def remove_ifexists(file, etime=sys.float_info.max):
     if os.path.exists(file):
         mtime = os.path.getmtime(file)
         if mtime < etime:
-            logger.info("Removing: {}", file)
+            logger.info(f"Removing: {file}")
             os.remove(file)
 
 
@@ -247,6 +247,7 @@ def get_steps_and_members_to_retrieve(
 
     steps = sorted(set(step_list))
     waitfor_steps = sorted(set(waitfor_list))
+    logger.debug("waitfor_steps: {}", waitfor_steps)
     # Get perturbed members only
     perturbed_members = [member for member in sorted(set(member_list)) if member != 0]
     # Construct dictionary with perturbed members and control depending on the
@@ -495,23 +496,43 @@ def move_files(
                 lockfile = f"{dest_file}.lock"
                 if os.path.exists(file_name):
                     shutil.move(file_name, dest_file)
-                if os.path.exists(lockfile):
+                if os.path.exists(dest_file) and os.path.exists(lockfile):
                     os.remove(lockfile)
+                    logger.info(f"Removed lockfile: {lockfile}")
+
+
+def wait4_file(file: str):
+    """Wait for a single file to appear in final location.
+
+    Args:
+        file (str): Path to file
+
+    Raises:
+        SystemExit: if lockfile disappears before file is created
+    """
+    lockfile = f"{file}.lock"
+    logger.info(f"Wait4 checking {file}")
+    while not os.path.exists(file):
+        if os.path.exists(lockfile):
+            logger.info(f"Waiting 10 sec. for {file}")
+            sleep(10)
+        else:
+            logger.info(f"No lockfile, stop waiting for {file}")
+            sleep(20)
+            if not os.path.exists(file):
+                raise SystemExit(f"Creation of {file} must have failed")
 
 
 def waitfor_files(
     tag: str, steps: List[int], members_dict: Dict[str, List[int]], target_dir: Path
 ):
-    """Wait for file to appear in final location.
+    """Wait for files to appear in final location.
 
     Args:
         tag (str): Name of tag
         steps (list): steps to wait for (processed elsewhere)
         members_dict (Dict[str, List[int]]): dict with members
         target_dir (Path): target directory where files should appear
-
-    Raises:
-        SystemExit: if lockfile disappears before file is created
     """
     for step in steps:
         for members in members_dict.values():
@@ -519,15 +540,8 @@ def waitfor_files(
                 # Define target file name. Member defaults to 0 if member is None
                 file_name = f"{tag}_{member or 0}+{step}"
                 dest_file = target_dir / file_name
-                lockfile = f"{dest_file}.lock"
-                while not os.path.exists(dest_file):
-                    if os.path.exists(lockfile):
-                        logger.info("Waiting 5 sec. for {}", dest_file)
-                        sleep(5)
-                    else:
-                        logger.info("No lockfile, stop waiting for {}", dest_file)
-                        sleep(30)
-                        raise SystemExit(f"Creation of {dest_file} must have failed")
+                wait4_file(dest_file)
+                logger.info(f"Waitfor, found: {dest_file}")
 
 
 def compile_target(
